ios: Re-support iOS 12 by removing translatable plurals in UI strings
This change is intended for a stable release atop v27.201, to fix a
pair of mistakes we made:

- When the UI started including FormatJS messages with this
  plural-handling syntax (35129e50e / cc53dd5a1), we didn't realize
  that, on iOS 12, they would render in a way that's not meant to
  face users, with characters like "}" in them. That's issue #5670.
- When we intended to de-support iOS 12 and 13 (783e56b23), we
  didn't actually do so; the App Store still offered the app for iOS
  12 and 13. That's issue #5672, fixed in `main` (e93d47f3d /
  PR #5673).

So, this release will be the last one that supports iOS 12 and 13,
and the text will be human-friendly for users stuck on iOS 12.

To keep it simple, just use English text for everybody. It would be
better to use even flawed translations for the messages, but it
would be hard to wrangle Transifex (or a Git merge) into giving us
the right mix of v27.201's translations for everything else and the
older translations for those two since-deleted strings "1 unread
message" and "{unreadCount} unread messages".

Fixes: #5670