ios: Revert iOS 12 bugfix from v27.202 release
This reverts commit f838a35f9, which fixed the iOS 12 bug #5670.
That bugfix involved a tradeoff: it disabled translation on three
user-facing messages (one being the prominent "X unreads" message in
`UnreadNotice`), replacing them with hard-coded strings in English.

That was an improvement for iOS 12 users, since they were affected
by #5670. But in the previous commit, we cherry-picked our commit
from `main` that enforced our intended iOS 14+ requirement. So,
while iOS 12 users can still get the #5670 fix in the v27.202
release (as the last release available to them), now we can make a
followup release v27.203 that puts the translations back, and won't
clobber iOS 12 users' installs with a version that has the #5670
bug.