internalLinks [nfc]: Dissolve `getLinkType` / `LinkType` indirection
This puts the logic to recognize a "stream link" right next to the
logic for building a stream narrow, and likewise for topic narrows,
etc.

Now it's quicker to read what `Narrow` we construct for a given
`url` param. We don't have to click through `getLinkType` and wonder
whether anything surprising is done at that layer.