msglist: Give error alert when link parses to Narrow we don't have data for
For PM narrows, we've been showing a really busted-looking screen
when we parse the link to a Narrow but we don't have data for the
user IDs; that's #5702.

For stream narrows, we've been taking the user out to a browser,
which is likely to be futile; that's #5698.

Fix both issues by, when we don't have data, preventing any
navigation and showing an error alert instead.

In the streams case (#5698), getNarrowFromLink wasn't distinguishing
the case of an unknown/nonexistent stream from the case of an
unparseable link. To change the behavior for the former but not the
latter case, we relax `parseStreamOperand` so that it still parses a
"new"-format operand even if it points to data we don't know about.
(Like we already do for PM operands.)

Fixes: #5698
Fixes: #5702