fetchActions tests: Don't spread value with opaque PerAccountState type
With Flow v0.175.1, which we'd like to upgrade to soon, we'd get
this error:

```
Error ---------------------------------------------------------------- src/message/__tests__/fetchActions-test.js:312:37

Cannot spread object literal because `PerAccountState` [1] is not an object. [not-an-object]

   src/message/__tests__/fetchActions-test.js:312:37
                                            v
   312|         const stateWithoutAccount = {
   313|           ...baseState,
   314|           accounts: [],
   315|         };
                ^

References:
   src/__tests__/lib/exampleData.js:684:18
   684| ): GlobalState & PerAccountState =>
                         ^^^^^^^^^^^^^^^ [1]
```

Probably to do with `PerAccountState` being opaque. (The `baseState`
value is returned by `eg.reduxStatePlus`, so it's a
`GlobalState & PerAccountState`.)

Fortunately it's easy to fix: this test is checking what happens
when a normal, boring message fetch happens when there isn't a
logged-in account to handle the payload. We can simulate that state
by making `state.accounts` empty, without touching other state
subtrees.