narrow [nfc]: Have specialNarrow take string-literals union, not `string`
Oddly, the regex was doing a looser check on the operand than
`specialNarrow` was: it was case-insensitive, and it was only
looking at the beginning of the string. That doesn't end up making a
difference, because `specialNarrow` would throw an error and we'd
return null, the same as if the regex hadn't matched the string.