internalLinks: Parse '/dm/…' and '/is/dm' in narrow links
These are synonymous with '/pm-with/…' and '/is/private',
respectively, and they'll start appearing in messages on newer
servers, as part of the renaming of "private message" to "direct
message".

Mobile still generates links with the old '/pm-with/…' format in
quote-and-reply. Those will still probably be supported
indefinitely, but it'd be cleanest to switch to generating the new
format, with a feature-level condition. Once the server/web-app
merge code that uses the new format, we'll know what feature level
to use.

Fixes: #5692