deps: Upgrade camera-roll to 5.2.4 and constrain to <5.3.0
5.3.0 says in the changelog that it adds support for the new React
Native architecture:
  https://github.com/react-native-cameraroll/react-native-cameraroll/releases/tag/v5.3.0

We're not yet ready to use the new architecture. Possibly the new
architecture doesn't also become *required* at the same time, but
we've had similar hopes in comparable situations and it doesn't
always work out that way.

Plus in 5.3.0 the TypeScript types start importing things with names
like TurboModule from the React Native types, and so to support that
in our Flow types we'd need changes to TsFlower. Leave that work for
later.