logging: Make `logging.info` log to Sentry
We have a log line

  NotifTroubleshootingScreen: MailComposer reports a sent email.

that I'd *thought* was getting sent to Sentry, but apparently not.
The line is meant to help us detect potential cases where the user
thought they'd contacted support through the app, but we don't see
an email on our side. We're not aware of such cases yet, but in
principle it could happen.

Since it seems helpful to have a function for logging to Sentry at
an "info" level, rewrite `logging.info` based on `logging.warn` and
`logging.error` so it does that. That should cause that
notif-troubleshooting line to go to Sentry.

The one other callsite of `logging.info` is redundant with a
breadcrumb added with the same information; remove it and an
unexplained test assertion about it.