deps: Upgrade to RN v0.68.7, from v0.68.5
Aimed at fixing #5729.

This small upgrade is nice and easy; the upgrade helper doesn't show
any other changes we need to make:
  https://react-native-community.github.io/upgrade-helper/?from=0.68.5&to=0.68.7

Release notes:
  https://github.com/facebook/react-native/releases/tag/v0.68.7

Incidentally we also get facebook/react-native@c3ad8ec7e in this
release, which makes the change we've been patching in with
`patch-package` to fix an Xcode build failure, so we can remove our
`patch-package` patch.

Tested on an iOS simulator (iPhone 14 Pro running iOS 16.4), and on
the office Android device (Samsung Galaxy S9 running Android 9). We
haven't managed to reproduce #5729 ourselves so I can't report an
observation of how this upgrade affects that issue. But upstream
discussion is promising, and from a quick look nothing seems to
break.

Fixes: #5729