ios [nfc]: Cut notifications library out of handling "initial notification"
And handle it ourselves, through our dedicated iOS native module
ZLPNotificationsStatus.

(The "initial notification" is a notification that, by being tapped,
launched the app from cold. PR #5676 was about notifications that
are tapped when the app is already open, and that PR fixed a known
bug.)

This, like #5676, makes our iOS notification handling more
transparent. My hope is to smooth the path to custom iOS
notification code in zulip-flutter (zulip-flutter#122, etc.) by
demonstrating a working implementation that's simple and avoids
depending on the details of
@react-native-community/push-notification-ios.

Cutting the library out of this codepath makes one small difference
in the payload we pass to `fromAPNs` on the JS side: the `aps`
property is now present. (So, remove a comment saying it might be
absent). We've never looked at that property, so this is NFC. Unlike
in 2ae6f798d / #5676, there wasn't a call to the library's custom
RCTConvert extensions to account for, as we did there with a
cautious call to RCTJSONClean.