android: targetSdkVersion to 33, this time handling new notifs permission
(cherry picked from commit 4fe472c38f39ba98dbfdc87859429065f28caac0)

We bumped this to 33 in 6f44474ac / v27.210, but then reverted it in
2e196e43d because of #5753, "Use the new POST_NOTIFICATIONS runtime
permission on Android".

Bump it again, this time with a fix for that issue: we declare the
new permission and add a runtime function call to request it in
context.

For users, this means:

- If on iOS: No change.

- If below Android 13: No change. A permission request follows the
  creation of a notification channel, so effectively it happens on
  app startup:
    https://developer.android.com/develop/ui/views/notifications/notification-permission#new-apps

- For a new app install on Android 13, the permission will now be
  requested in context instead of at startup:
    https://developer.android.com/develop/ui/views/notifications/notification-permission#new-apps

- For existing app installations when Android is upgraded to 13 or
  later, the permission should automatically be granted without a
  prompt as long as notifications hadn't been explicitly disabled
  before the upgrade. This should still work across a "backup and
  restore":
    https://developer.android.com/develop/ui/views/notifications/notification-permission#existing-apps
  (However: we are tracking #5484 as an unrelated cause of
  notifications breaking across a "backup and restore".)

If RN and Android decide it's helpful, a "rationale" is shown to the
user just before the new in-context permission request. That's a
modal showing a message we provide (see added code), with an OK
button to close the modal and proceed to Android's plain Yes/No
dialog for the permission request.

Fixes: #5753