icons: Add the Zulip "follow" icon
This SVG file is copied verbatim from the web app.  Ideally we'd
do that by upgrading `@zulip/shared`... but now that this codebase
is legacy and we're spending less time maintaining it, we've
accumulated a backlog of changes there that are incompatible and
require adjustments in our code.  So the expedient thing is to
just take the individual icon we currently need and pull it into
our tree directly.

Moreover, we want this icon inside the webview.  This makes only
the second Zulip icon we've put inside the webview, after the
"play_button" icon we added just the other day; and I wouldn't say
we've developed a good streamlined pattern for it.  But for this
legacy codebase, we'll content ourselves with just following the
pattern set by "play_button".

That also means duplicating the icon's SVG file into two places.
Awkward, but at least it's not any kind of space-consumption issue:
the file is under a kilobyte in size.