test: Fix RN Jest setup to work in jsdom on recent Node releases
This patch is a backport of commit facebook/react-native@cf631ad59
from upstream.  I've also pushed it to a branch in our fork:
  https://github.com/zulip/react-native/compare/v0.68.7...0.68.7-zulip

In particular this avoids a symptom which I otherwise see
on Node v18.16.1:

```
 FAIL  src/webview/__tests__/generateInboundEventEditSequence-test.js
  ● Test suite failed to run

    TypeError: Cannot redefine property: performance

      at Object.<anonymous> (node_modules/react-native/jest/setup.js:20:20)
```

The symptom affects three of our test files:

    src/__tests__/test-test.js
    src/webview/__tests__/generateInboundEventEditSequence-test.js
    src/webview/js/__tests__/rewriteHtml-test.js

because those are the three that have `@jest-environment` annotations
calling for "jsdom" or related environments.