notifs: Revise "server not set up" message; drop condition on Role.Admin
One of the Role.Admin conditions had actually been going the wrong
way, showing the admin message to non-admins and the non-admin
message to admins. Fortunately that was on the notification
troubleshooting screen, which you can't have reached unless the
server had been set up for notifications earlier in the mobile app's
session.

Fixes-partly: #5785