api: Intentionally handle removal of realm_email_address_visibility
I can't find a message in `#api design` on CZO that's explicitly
about removing realm_email_address_visibility from the register
response. The `#api design` discussion about transitioning that from
an org- to a user-level setting is here:
  https://chat.zulip.org/#narrow/stream/378-api-design/topic/email.20address.20visibility/near/1296131
That discussion includes mentions of the user-object fields that are
about email addresses, and it includes mentions of the per-user
visibility setting, but I don't see anything about removing the
realm setting from the /register response.

In retrospect, its removal there would have been an obvious thing to
anticipate and handle in our types. Oh well, better late than never.

Luckily I don't think there was a user-facing bug here. At FL 163+,
the value that we've been typing as the enum EmailAddressVisibility
will actually be `undefined`, but that's OK because on 163+ we don't
end up doing anything with it.