AvatarItem: Remove animation
We've so far hesitated to convert this to a function component
because of its animation logic, which would take a bit of thought to
convert correctly. But we want to use `useSelector`, and for that it
needs to be a function component. So, get rid of the animation. A
small sacrifice in this page, which has really long loading times
and pretty extreme jank when you tap a user.