android build: Update jsBundleDirRelease for AGP 7.1+, fixing crash on startup
This is a new incarnation of the same underlying React Native
issue seen in 8f8a266af and 833d083a9, back in 2019 and 2018
respectively.  RN hard-codes what intermediate directory it
expects the Android Gradle Plugin to look for assets in; but that
directory isn't any kind of stable API, and in fact it changes
from time to time in new versions of AGP.  It changed again
in AGP 7.1.0.

More recent versions of RN know about this latest change (with yet
another hardcoding, conditioned on a new range of AGP versions):
  https://github.com/facebook/react-native/pull/33057
But ours doesn't, because we haven't been updating RN for a while in
this legacy codebase.  So just update the directory we force it to.

See discussion:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/release.20build.20missing.20JS.20bundle/near/1746718

[greg: expanded commit message]