deps: Upgrade @react-native-async-storage/async-storage to 1.23.1 (latest)
Changelog:
  https://github.com/react-native-async-storage/async-storage/releases

This new version has a "privacy manifest", which should help us
toward #5847.

For why this upgrade is unlikely to break things, see:
  https://github.com/zulip/zulip-mobile/pull/5309#discussion_r847620787

For why the version was constrained with ~ instead of ^ see
47b7eadfa, an Expo SDK upgrade:

  [T]he `expo upgrade` command
  went through a phase it described as "Updating packages to
  compatible versions (where known)," and that changed several of our
  dependency ranges, even of non-Expo things.

It would actually be better to remove the dependency entirely. It's
our legacy way of storing data on the device, and we have a
migration to the new way that's now been active for over two years
(see caf3bf999). The proportion of our users who still benefit from
what this dependency offers will be vanishingly small at this point.

But I'm not totally comfortable removing it without more careful
thought than we have time for right now in this legacy codebase.
(One step we'd want to consider, if doing that, is adding and
releasing a migration to clear out the contents of the legacy
storage.)

Tested on my iPhone and on the office Android device, on both a
first and a subsequent startup of the app. No crashes or errors
observed.

Related: #5847