deps: Upgrade react-native-simple-toast to 3.3.1, the latest
Unfortunately this causes a peer-dependency warning:

  warning " > react-native-simple-toast@3.3.1" has incorrect peer
    dependency "react-native@>=0.71.0".

but no issues are observed in manual testing of this library's
functionality on iOS or Android.

While we're at it, use TsFlower for the types. The types tell us
that `Toast.show`'s duration parameter is required; so, include it,
even though the implementation has logic to fall back to Toast.SHORT
if the argument isn't passed.

This lets us pull in the latest version of the "Toast" pod for iOS;
that's version 4.1.1. That version has a "privacy manifest" --

  scalessec/Toast@401141d79
  scalessec/Toast@607b12954
  scalessec/Toast@85b492dd8

-- which should hopefully help us with #5847. Granted, the Toast
pod's privacy manifest file seems pretty boring; I think it just
declares that the library doesn't access any sensitive data. Still,
"Toast" is explicitly included in a list published by Apple of
"commonly used SDKs" that require their own privacy manifest:
  https://developer.apple.com/support/third-party-SDK-requirements/
(In fact, I think it's the only one of zulip-mobile's dependencies
that appears in that list.)

Related: #5847