outboxActions: Show dialog if message send fails with informative ApiError
This is our minimal support for #5870 in this legacy codebase.
Supporting it properly with a client-side check of the setting is
more effort than we can spare here, because it requires implementing
the group-based permissions system.

Probably more error handling is called for in general (like for
network or server issues), but #3881 ("Sending outbox messages is
fraught with issues") is complicated and it's probably best to leave
it be.

Fixes: #5870