android: Build RN from source, at our fork, for targetSdkVersion 34
This gets us zulip/react-native@5c36f102f and its fixup commit
zulip/react-native@b7b2f6c22 (on the `0.68.7-zulip` branch), to fix
a build failure with `targetSdkVersion = 34` without having to
upgrade to RN v0.73. (The build failure is in debug builds only, but
we still need to make those to develop fixes in this legacy project
as needed.)

Thanks to Greg for investigating the build failure:
  https://github.com/zulip/zulip-mobile/issues/5877#issuecomment-2364347383

Since (as part of working on this) the `0.68.7-zulip` branch also
includes the things that were in the patch-package patch, that patch
becomes redundant and we delete it here.

Done by digesting the history of RN's doc on building RN from
source:
  https://github.com/facebook/react-native-website/commit/22da0cc63
and making some simplifications.

The line `ndkVersion = 24…` sets a variable that ReactAndroid reads,
to control the NDK version it builds with. It would otherwise build
with NDK 21, which is old and chosen automatically based on
ReactAndroid's targetSdkVersion of 31. We need it to be at least 24
for Apple Silicon support.