android build: Allow Gradle to use twice as much memory
With the upcoming change to build RN from source, CI was failing
with out-of-memory errors. This should help with those.

Why twice as much memory? See 367778858, which bumped this value
from 1024 to 1280. From the commit message there:

  Compare facebook/react-native@a2a703247, which happened on the way
  to RN v0.68.  That one had the explanation

    > Updates maximum heap size for the gradle build to account for
    > building RN from source when new architecture is enabled.

  which doesn't apply as we aren't yet adopting RN's New
  Architecture. But perhaps if we did, we'd need to increase the
  heap size even more: that commit takes it to 2 GiB.

We're not adopting RN's New Architecture, and we don't plan to in
this legacy codebase. But we will be building RN from source, so
this seems like a good value to try.